First commit of Geo app.
Currently Geo finds your location when you click On buton and writes it to the database. However, it writes new entries every time you click on, so the findOne function stops working and it won't display correctly.
I need to make it so that if the entry exists, it updates the value instead.

The user functionality isn't added yet, so the database is populated with dummy info.