After spliting shit into services, I finally got the service to work.
The issue was that the multiple angular.module('geoApp') calls were fucking each other up, but it seemed like the geoApp that was supposed to create the module was getting called after the one that was calling the module in order to create the service.
It was all fucked.

SO now I made the service a service off the dist component's app ('dist'), so that it could be called in the same file. So it's working now, despite the fact that it seems to me that it should have worked before...