Also pretty intense style changes. Made it so that the text is all aligned at the bottom of each entry. Made checkbox bigger. Learned that the align-content and shit only applies to the elements within the flex element you're styling.
Made it so that when you scroll all the way to the bottom of the list, there's a margin at the bottom that will ensure that you can see the last entry above the button bar.

Made all the database calls (except for the helper functions in the view) that use methods "secure" by adding in logic to throw an error if the user data being accessed isn't the user that's logged in.